Description
Objective - 
Statistically analyzing ongoing trends in the given dataset of vaccination rates in the California School Districts and recommending the government about steps to be taken to increase the total vaccinations.

Contributions - 
a) Exploratory analyses of dataset of 1000 observations (sample of original dataset) with histograms, bar-plots, and boxplots/violin plots to understand data distribution and then performed data cleaning. Identified the past trends in the vaccine rate variations.
b) Implemented the Descriptive and Inferential statistical results to report the analyses with 95% accuracy.
c) Incorporated unbiased data along with data transformation to understand the underlying correlation between dependent (vaccination rates) and independent variables.
d) Consolidated and presented the statistical analyses report confirming the results obtained with numerous tests and recommended important factors to be focused on to increase the total vaccinations in California Districts which are lacking.